Fix PHPStan level 5 errors: type casting, parameter bugs, PHPDoc fixes
- Fix ActivityTracker::track() called with 5 params instead of 4
- Fix potential division by zero in ProfileShortcode
- Add type casts for str_repeat, base_convert, wp_get_attachment_image
- Fix PHPDoc return types for WP_Error in CSV and ListsEndpoint
- Configure phpstan.neon with WordPress-specific ignore patterns

+++ src/API/BadgeEndpoint.php
@@ -224,7 +224,7 @@ class BadgeEndpoint {
 			: $business_name;
 
 		// Generate stars SVG.
-		$stars_svg = self::generate_stars_svg( $rating, $dim['font'] * 1.2 );
+		$stars_svg = self::generate_stars_svg( $rating, (int) ( $dim['font'] * 1.2 ) );
 
 		// Site name.
 		$site_name = get_bloginfo( 'name' );
